python/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
google-cloud-aiplatform
vertexai
requests
httpx
python-dotenv